# internal/adapters/http/openapi.yaml
...
openapi: 3.0.3
info:
  title: situs API
  description: >-
    Lokaler, schreibgeschützter Dienst für EUNIS-Habitattypen: Art →
    Habitattypen (mit Rolle) und Habitattyp → Arten / Syntaxa / Crosswalks,
    aus gepinnten EEA- und ESy-Artefakten, mit deutschen Labels und einem
    Anhang-I-Crosswalk (annex1).
  version: 0.3.0 # x-release-please-version
  license:
    name: MIT
servers:
  # Relativ, absichtlich: Swagger-UI baut die "Try it out"-Anfragen aus diesem
  # Eintrag. Eine absolute URL mit festem Port zeigt am tatsächlich benutzten
  # Port vorbei, sobald der Dienst nicht auf dem dort genannten läuft. Relativ
  # löst gegen die Herkunft der Seite auf und trifft damit immer den Server, der
  # sie ausgeliefert hat.
  - url: /
    description: Derselbe Server, der diese Spezifikation ausliefert
paths:
...
  /v1/habitat-type/{typology}/{code}:
    get:
      summary: Ein Habitattyp mit Arten, Syntaxa und Crosswalks
      description: >-
        Dieselbe Route trägt EUNIS und Anhang I — ein Habitattyp wird immer über
        (typology, code) adressiert. Die Crosswalk-Liste trägt beides: andere
        EUNIS-Fassungen und Anhang-I-Entsprechungen. Sie ist leer, wenn es keine
        gibt; bei Anhang I ist das der Normalfall.
      operationId: habitatType
      parameters:
        - $ref: "#/components/parameters/Typology"
        - $ref: "#/components/parameters/Code"
        - $ref: "#/components/parameters/Lang"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      responses:
        "200":
          description: Der Habitattyp.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitatTypeDetail"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/habitat-type/{typology}/{code}/species:
    get:
      summary: Artenliste eines Habitattyps, optional nach Rolle gefiltert
      operationId: habitatTypeSpecies
      parameters:
        - $ref: "#/components/parameters/Typology"
        - $ref: "#/components/parameters/Code"
        - name: role
          in: query
          required: false
          description: Filtert auf eine Rolle; ohne Angabe werden alle geliefert.
          schema:
            $ref: "#/components/schemas/Role"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      responses:
        "200":
          description: Die Arten des Habitattyps.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpeciesEntry"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/syntaxon/{id}/habitat-types:
    get:
      summary: Habitattypen einer Pflanzengesellschaft
      description: >-
        Zeigt das m:n — dieselbe Gesellschaft kann in mehreren Habitattypen
        vorkommen.
      operationId: syntaxonHabitatTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Die verknüpften Habitattypen.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitatTypeSummary"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Typology:
      name: typology
      in: path
      required: true
      description: >-
        Typologie in ihrer Fassung. Ein leeres Segment fällt auf eunis@2021
        zurück (der Pfad selbst verlangt ein Segment, weglassen lässt sich es
        also nicht). Eine unbekannte Typologie ist INVALID_QUERY, kein
        NOT_FOUND.
      schema:
        type: string
        example: eunis@2021
    Code:
      name: code
      in: path
      required: true
      description: Code des Habitattyps innerhalb der Typologie.
      schema:
        type: string
        example: R22
    Lang:
      name: lang
      in: query
      required: false
      description: >-
        Antwortsprache, alternativ per Accept-Language. Lokalisierte Felder
        kommen additiv hinzu — name_en bleibt immer gesetzt. Default ist en.
      schema:
        type: string
        enum: [en, de]
    Area:
      name: area
      in: query
      required: false
      description: >-
        Gebietscode nach WGSRPD Level 3 — das Frontend leitet ihn aus GPS ab, so
        braucht situs keine ISO-Zuordnung. Ist er gesetzt, tragen die Arten- und
        Habitattyp-Einträge zusätzlich in_area. Ein Code, zu dem der Index keine
        Verbreitungsdaten hat, ist INVALID_QUERY und nicht "kommt nicht vor" —
        sonst sähen ein Tippfehler und eine echte Abwesenheit gleich aus.
      schema:
        type: string
        example: GER
    OnlyInArea:
      name: only_in_area
      in: query
      required: false
      description: >-
        Verwirft die sicheren Abwesenheiten (in_area=false). Was nicht beurteilbar
        ist (in_area fehlt), bleibt immer stehen. Ohne area ist der Parameter
        absichtlich wirkungslos statt ein Fehler. Auf dem Batch-Endpunkt markiert
        area nur, verwirft aber keinen Eintrag: dort gehört response[i] zu
        concept_ids[i], ein Eintrag darf also nicht fehlen.
      schema:
        type: boolean
  responses:
    InvalidQuery:
      description: Die Anfrage ist fehlerhaft (INVALID_QUERY).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Es gibt keinen solchen Eintrag (NOT_FOUND).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # Dieselbe Fehlerhülle wie überall. Sie ist nur an /v1/info ausgewiesen,
    # weil dort ein Auslöser benannt und getestet ist (eine fehlschlagende
    # Index-Abfrage); die anderen Routen können INTERNAL_ERROR ebenfalls
    # liefern, aber nur als unspezifizierten Ausfall.
...
  schemas:
    Role:
      type: string
      enum: [diagnostic, constant, dominant]
    # Fehlende Angaben sind abwesende Felder, nie ein Platzhalter oder eine
    # substituierte Null: level fehlt bei Typologien ohne Hierarchie, priority
    # gibt es nur bei annex1, name_de nur wenn ein Label existiert.
    # name_de ist ein Objekt, nicht ein String mit Geschwisterfeld: ein Client soll
    # value nicht lesen koennen, ohne provenance daneben zu sehen. Manche Labels
    # sind amtlich, andere hat situs selbst uebersetzt — das darf nicht abtrennbar
    # sein.
    GermanLabel:
      type: object
      required: [value, provenance, source]
      properties:
        value:
          type: string
        vernacular:
          type: string
          description: >-
            Etablierter deutscher Begriff. Fehlt, wo keiner denselben Umfang
            traegt wie der Typ — ein leerer Wert waere eine Behauptung.
        provenance:
          type: string
          enum: [official, curated, derived, situs]
          description: >-
            situs bedeutet: situs hat uebersetzt, keine externe Quelle steht
            dahinter.
        source:
          type: string
    HabitatTypeSummary:
      type: object
      required: [typology, code, name_en]
      properties:
        typology:
          type: string
          example: eunis@2021
        code:
          type: string
          example: R22
        level:
          type: integer
        name_en:
          type: string
        name_de:
          $ref: '#/components/schemas/GermanLabel'
        priority:
          type: boolean
    SyntaxonRef:
      type: object
      required: [id, rank, name]
      properties:
        id:
          type: string
        rank:
          type: string
          enum: [class, order, alliance]
        name:
          type: string
        author:
          type: string
          description: >-
            Autorschafts-Zitat, direkt aus FloraVeg.EU übernommen; fehlt, wenn
            kein FloraVeg-Treffer gefunden wurde.
        parent_id:
          type: string
          description: >-
            Verweist auf einen FloraVeg-Klasse-/Ordnungscode (reine
            String-Referenz, kein Fremdschlüssel); fehlt, wenn unbekannt.
...
    CrosswalkRef:
      description: >-
        Die Gegenseite einer Entsprechung, gesehen vom abgefragten Typ. Der
        Qualifier liest sich immer als "abgefragter Typ <qualifier> dieser Typ".
      type: object
      required: [typology, code, qualifier]
      properties:
        typology:
          type: string
        code:
          type: string
        qualifier:
          type: string
          example: "="
    AggregateSource:
      description: >-
        name is absent: the index only persists the aggregate's concept id,
        never its name.
      type: object
      required: [concept_id]
      properties:
        concept_id:
          type: string
        name:
          type: string
    SpeciesEntry:
      description: >-
        verbatim_name ist immer gesetzt; concept_id fehlt, wenn der Name nicht
        auf ein Konzept auflösbar war — solche Zeilen werden behalten.
        provenance/derived_from erscheinen nur bei einer aus einer Sammelart
        abgeleiteten Zeile.
      type: object
      required: [verbatim_name, role]
      properties:
        concept_id:
          type: string
        verbatim_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        fidelity:
          type: number
        constancy:
          type: number
        in_area:
          $ref: "#/components/schemas/InArea"
        provenance:
          type: string
          enum: [derived_from_aggregate]
        derived_from:
          $ref: "#/components/schemas/AggregateSource"
...
    HabitatTypeDetail:
      allOf:
        - $ref: "#/components/schemas/HabitatTypeSummary"
        - type: object
          required: [species, syntaxa, crosswalks]
          properties:
            species:
              description: >-
                Nach Rolle gruppiert; die drei bekannten Rollen sind immer
                vorhanden, ggf. als leere Liste.
              type: object
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/SpeciesEntry"
            syntaxa:
              type: array
              items:
                $ref: "#/components/schemas/SyntaxonRef"
            crosswalks:
              type: array
              items:
                $ref: "#/components/schemas/CrosswalkRef"
...
    InArea:
      description: >-
        Dreiwertig: true/false, wenn der Index Verbreitungsdaten zum Konzept
        hat, und abwesend, wenn die Frage nicht beantwortbar ist (keine
        Konzept-ID oder keine Verbreitungszeilen). Ohne ?area= fehlt das Feld
        immer — Abwesenheit ist hier "unbekannt", nie "kommt nicht vor".

        Auf POST /v1/species/habitat-types trägt nur ConceptResolution selbst
        dieses Feld: in den verschachtelten habitat_types-Einträgen fehlt es
        auch mit ?area=, denn die Aussage gilt der Art und wäre dort für jeden
        Habitattyp derselben Art dieselbe Konstante. Auf
        GET /v1/species/{conceptId}/habitat-types, wo es keine Eintragsebene
        darüber gibt, ist habitat_types[].in_area dagegen gesetzt.
      type: boolean
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string